cosf/sinf + static_cast<uint16_t>

+++ AtlasDeconstructor/src/AtlasDeconstructor.cpp
@@ -49,12 +49,12 @@ AtlasDeconstructor::getPatchIdMap(const std::vector<Vec2i> &atlasSize,
   for (const auto &sz : atlasSize)
   {
 	  PatchIdMap patchMap(sz.x(), sz.y());
-	  std::fill(patchMap.getPlane(0).begin(), patchMap.getPlane(0).end(), 65535);
+	  std::fill(patchMap.getPlane(0).begin(), patchMap.getPlane(0).end(), static_cast<uint16_t>(65535));
 	  patchMapList.push_back(std::move(patchMap));
   }
 
   for (auto id = 0u; id < patchList.size(); id++)
-    writePatchIdInMap(patchList[id], patchMapList, id);
+    writePatchIdInMap(patchList[id], patchMapList, static_cast<uint16_t>(id));
 
   return patchMapList;
 }
